requirements.txt
streamlit==1.14.0
tensorflow==2.16.1
numpy==1.23.4
Pillow==9.3.0
tensorflow-hub>=0.16.1

